CTFs/Ethernaut/Fallback/pyproject.toml
[tool.poetry]
name = "fallback"
version = "0.1.0"
description = "ape solution for ethernaut fallback"
authors = ["SAIKAT KARMAKAR <aviksaikat007@protonmain.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-ape = "^0.6.8"
ape-alchemy = "^0.6.1"
ape-ganache = "^0.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
